+++ Info.go
@@ -5,7 +5,7 @@ package uymas
 // @Name:    Uymas 包信息
 
 const (
-	Version = "0.2.0"
+	Version = "0.2.1-alpha"
 	Release = "20181106"
 	Author  = "Joshua Conero"
 	Name    = "uymas"